Initial Commit
The initial commit of a plugin to better detect LPC files using
the .c extension. This plugin negates the need to set the
boolean defined in vim :help lpc.

It is based on the check in $VIMRUNTIME/filetype.vim and checks
the first 500 lines of the .c file as opposed to the first 12.